requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7

# HTTP and API
requests==2.31.0

# Database
# sqlite3 is included in Python standard library

# Utilities
python-dotenv==1.0.0

# Logging (optional but recommended)
colorlog==6.8.0